refactor(agent-info): rename agents-info to agent-info for consistent naming and improved clarity in imports
