+++ plugins/main/public/components/common/welcome/agent-info/agent-info.tsx
